requirements.txt
# HKEX → Notion tool
playwright>=1.49.0,<2
beautifulsoup4>=4.12.0,<5
requests>=2.31.0,<3
python-dotenv>=1.0.0,<2
PyYAML>=6.0.0,<7
pdfplumber>=0.11.0,<0.12
lxml>=5.0.0,<6
